feat: move connect mobile into session section
